fix(cdp): fix macOS build and lint findings on the arm64 path
- Move launchBrowser into OS-split files (launch_linux.go / launch_darwin.go)
  so the shared player.go no longer references the linux-only
  chromiumProfileDir. This fixes the darwin build (Apple Silicon is
  darwin/arm64, which previously took the persistent-profile branch and failed
  to compile), and keeps the persistent-profile + warm-up logic correctly
  scoped to linux/arm64.
- Lowercase two error strings to satisfy staticcheck ST1005.
- Annotate the discovered-browser os.Stat with nolint:gosec (G703): the path is
  operator-provided (VIBEZ_CHROME_PATH) or a fixed system location.
